Add self-delete protection

+++ src/main/java/com/todo/demo/service/MemberService.java
@@ -5,10 +5,12 @@ import com.todo.demo.entity.TodoItemEntity;
 import com.todo.demo.repository.MemberRepository;
 import com.todo.demo.repository.TodoItemRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Service;
-import org.springframework.validation.AbstractBindingResult;
 
+import javax.swing.*;
 import java.util.List;
 
 @Service
@@ -24,7 +26,7 @@ public class MemberService {
     public List<Member> findAll(){
         return memberRepository.findAll();
     }
-
+    //TODO Add password change
     public Member save(Member member){
         BCryptPasswordEncoder encoder = new BCryptPasswordEncoder();
         member.setPassword(encoder.encode(member.getPassword()));
@@ -32,11 +34,15 @@ public class MemberService {
         return member;
     }
     public void deleteMember(String username){
-        List<TodoItemEntity> itemList = todoItemRepository.findAllByOwner(username);
-        for (TodoItemEntity a: itemList) {
-            itemService.delete(a.getId());
-        }
-        Member member = memberRepository.findMemberByUsername(username);
-        memberRepository.delete(member);
+        //TODO Add self-delete prevent
+        /*Authentication principal = SecurityContextHolder.getContext().getAuthentication();
+        if (!principal.getName().equals(username)){*/
+            List<TodoItemEntity> itemList = todoItemRepository.findAllByOwner(username);
+            for (TodoItemEntity a: itemList) {
+                itemService.delete(a.getId());
+            }
+            Member member = memberRepository.findMemberByUsername(username);
+            memberRepository.delete(member);
+        //}
     }
 }
